First commit of Future and Promises.
Continuing on the goals of javaslang, the Future api is modeled on the
Scala Future api, but is focused on Java compatibility.

Currently it's based upon Java's CompletableFuture internally, though it
might make more sense later to cut that out and handle callbacks
ourselves.

Some features are still very speculative, like the
java.util.concurent.Future compatibility. Method names, comments and
other things are very open for discussion.

Futures, the helper class for converting a Iterable or Tuple of Futures
into a Future of an Iterable or Tuple, is not done yet, and I was
planing on putting it in the Generator to handle the Tuples. Some of the
code is already there, but I was having some trouble running the
Generator and decided to commit what I have now and worry about that
later. It's not active at the moment.

FutureTest is not even close to exhaustive and needs more work.