Reworked Iterable sequence
Forget the Try's, it's not really the right way to do it. Instead the
result will fail if even a single component failed, just like the
Tuples. Doing so means there is no need for the flatSequence methods, so
they have been removed. This should be much simpler and intuitive.