Modified callBacks to ensure polling.
Tested the callBacks more carefully, found issues and fixed them.

Added code to the java.util.concurrent.Future converter to make sure
that if it's a CompletableFuture in disguise that it's treated like one
for efficiency's sake.

Also ran code formatter, resulting in lots of worthless noise in the
commit log.