Fixed FutureTest, issue with Future
Made FutureTest use Java's CompletableFuture to help test, with the
assumption that it will always work properly.

Added more tests to get 100% code coverage of Future and Promise. Tests
likely aren't exhaustive.

Discovered a situation where a callback was being missed. Not exactly
sure how, but it was there for sure. So I've moved to a slightly
different method of synchronizing on the callback queue to ensure that
this never happens, and I think this one should be fool-proof.