Moved donation button to the bottom. (#1887)
Donation is a good thing. If people are willing to spend a bit, it could compensates our costs (mainly coffee ;) but also domain/hosting, hardware to develop, maybe some marketing, ...). But currently only one person(!) gave back. (Thank you!) Our current balance is 10 EUR.

However, I think donation is not the right thing to move Javaslang to the next level. If we plan to have someone work full-time on it there must be another way to get capital. Maybe Kickstarter is the right way. We will see...