Replace Validation class with one that allows for nested validation of deep Pojos.
Using the current Validation class for anything other than a flat classes leads to unreadable code.

For example, I wish to parse and validate a Person object, which contains a name String and an Address object. Trying to combine these validations leads to very unreadable code. The chief problem is validation of a single field has a single error, and a combined validation has a sequence of errors. These cannot be combined in a fluent way.

By ensuring that every validation returns a sequence of errors, I can combine any validations I want fluently. This allows for complex validation rules with nested validation.

I have deprecated the current Validation class and created a new one in another package. As some of the deprecation warnings for the generated code were appearing despite @SuppressWarnings("deprecation") being switched on (a known issue), I have had to disable deprecation warnings.

The library is still backward compatible, and it's pretty simple to upgrade your deprecated classes and methods with the new ones. Any feedback is appreciated.