Replace Validation class with one that allows for nested validation (#2073)
* Replace Validation class with one that allows for nested validation of deep Pojos.

Using the current Validation class for anything other than a flat classes leads to unreadable code.

For example, I wish to parse and validate a Person object, which contains a name String and an Address object. Trying to combine these validations leads to very unreadable code. The chief problem is validation of a single field has a single error, and a combined validation has a sequence of errors. These cannot be combined in a fluent way.

By ensuring that every validation returns a sequence of errors, I can combine any validations I want fluently. This allows for complex validation rules with nested validation.

I have deprecated the current Validation class and created a new one in another package. As some of the deprecation warnings for the generated code were appearing despite @SuppressWarnings("deprecation") being switched on (a known issue), I have had to disable deprecation warnings.

The library is still backward compatible, and it's pretty simple to upgrade your deprecated classes and methods with the new ones. Any feedback is appreciated.

* Increase code coverage.

* Add package-info.java file.

* Remove new 'fluent' package and replace the existing Validation.
This is because vavr 1.0 doesn't need to be backward compatible, so no need to duplicate and deprecate classes.

* Fix type checking.

* Remove methods that are no longer supported with vavr 1.0.

* Fix a test.

* Possibly fix a test that passes locally but fails on the pipeline.

* Change generic type of the sequence of errors.

* Change generic type of the sequence of errors.
Fix some typos.

* Merge remote-tracking branch 'upstream/master'

# Conflicts:
#	vavr/src/main/java/io/vavr/Value.java
#	vavr/src/main/java/io/vavr/control/Validation.java
#	vavr/src/test/java/io/vavr/MatchTest.java