Streamline code generation
This commit starts modernizing the Gradle build by introducig a proper
task type for code generation. It makes use of the build cache and ensures
that the task declares its inputs/outputs properly.

It also downloads the Ammonite toolchain and caches it in the Gradle user
home, meaning that even if you delete the build directories or checkout
in a different directory, the toolchain wouldn't be re-downloaded.

A proper extension for configuring the Ammonite toolchain has been declared:

```
ammonite {
   ammoniteVersion = '...'
   scalaVersion = '...'
}
```

Because I'm not familiar with Ammonite I didn't know how to properly
pass the configred output directory for generated sources to the script,
which would be a good improvement vs hardcoding as it is today.

The build script also makes use of the toolchain support to make sure
that the JDK being used for compiling is JDK 8, independently of what
is used to run Gradle.

Explicit dependencies between tasks have been replaced wih implicit
dependencies, leveraging the lazy configuration API.

Because the code generation step is cached, it's pretty cheap, so the
`src-gen` directory was removed from source control: checking in generated
sources is considered a bad practice in general.