Migrate to the `maven-publish` plugin
The `maven` plugin has been deprecated for some time already and Gradle 7
will remove it. This commit migrates the publishing mechanism to the new
publish plugin.

It also configured the Nexus Staging repository plugin in order to simplify
publication to Maven Central and reduce configuration.

Note: I wasn't able to check the interaction with the "release" plugin, but
I could check that running `./gradlew publishToSonatype closeAndReleaseSonatypeStagingRepository`
does the expected work.