Fixing Methodref contant errors by changing Option back to an interface
I encountered many `Method 'io.vavr.control.Option io.vavr.control.Option.none()' must be Methodref constant` errors when trying to use the current alpha version of vavr with a current version of spring boot.
I'm not quite sure what this error means but changing it the abstract class back to an interface seemed to solve this problem.
I guess there is a reason why you changed it to an abstract class. If there is a better way to solve this I'm interested in how this can be done.